Move identifier-rules to common location
